feat: embed sprout CLI in dev-mcp as a shim, add NIP-OA auth tag support
Expose sprout-cli as a library crate and embed it in sprout-dev-mcp
using the same multi-call binary pattern as rg and tree. When the
agent runs `sprout list-channels` through the shell tool, the dev-mcp
binary handles it in-process — no external binary needed.

Also adds NIP-OA auth tag support to the CLI, bringing it to parity
with sprout-mcp: SPROUT_AUTH_TAG is parsed, Schnorr-verified at
startup, injected into every signed event via SproutClient::sign_event(),
and sent as an x-auth-tag HTTP header on all REST requests.

sprout-cli remains independently usable as a standalone binary.
The --auth-tag / SPROUT_AUTH_TAG flag is optional and backwards
compatible — when absent, behavior is unchanged.
