+++ crates/sprout-dev-mcp/src/shim.rs
@@ -20,6 +20,9 @@ impl Shim {
         let tree_link = dir.path().join("tree");
         symlink(&self_exe, &tree_link)?;
 
+        let sprout_link = dir.path().join("sprout");
+        symlink(&self_exe, &sprout_link)?;
+
         let original = std::env::var_os("PATH").unwrap_or_default();
         let mut new_path = std::ffi::OsString::from(dir.path());
         if !original.is_empty() {
